requirements.txt
flask
flask-cors
google-ai-generativelanguage
google-generativeai
python-dotenv
firebase-admin